Set variables to expressions instead of values
Previously, to set a variable x to 5, the syntax was:
(set x 5)
Now the syntax is:
(set x (lit 5))
This chnage was introduced to make it easier to set variables to
non-literals (which was not implemented yet). For example to set x to y:
(set x (get y))

The call syntax was also updated from:
(call f (arg1_expr) (arg2_expr) ...) to
(call f : (arg1_expr) (arg2_expr) ...)
because there was a parsing issue when there were no arguments and no
space between the function name and the closing parenthesis. This is
not a problem with the new syntax, which for a function with no
arguments would be:
(call f :)

Indentation of the output cpp code was also added.