Copy the correct number of list elements to device
The previous code used the number of iterations in a parallelized loop
to determine a guess at how many elements should be copied from host
to device when a list was passed, but this guess could easilty be
incorrect. This update makes the host to device copy actually copy the
correct number of elements (the entire list, but no more).