Only setup cuda files if parallelizing
The .cu and .cuh files were previously always setup and incorporated
into the Makefile, even when the code was not parallelized and so the
cuda files were mostly empty and did not do anything. On some systems
this prevented users from running non-parallelized code if they did not
have CUDA installed because they lacked the <cuda_runtime.h> file. Now
users without CUDA can run the code as long as nothing is parallelized.