Add `seq_loop' expression to force sequential loop
Sequential loops can be beneficial for testing the correctness of a
parallelized equivalent, and when the overhead of copying data to and
from the device outweighs the benefit of parallelization. Previously,
sequential loops were only possible by making the loop not
parallelizable, but this commit adds the `seq_loop' expression so the
user can explicitly tell the program not to parallelize a loop.